// src/app/model/iTune-file-type.enum.ts
export enum iTuneFileType {
  book = 'book',
  album = 'album',
  audio = 'coached-audio',
  movie = 'feature-movie',
  booklet = 'interactive- booklet',
  video = 'music-video',
  pdf = 'pdf',
  podcast = 'podcast-episode',
  software = 'software-package',
  song = 'song',
  episode = 'tv- episode'
}